Move demo task trigger to demo app and update docs
