+++ demo/templates/demo/demo_home.html
@@ -142,7 +142,7 @@
             submitBtn.innerHTML = '<span class="spinner-border spinner-border-sm" role="status"></span> Starting...';
             
             try {
-                const response = await fetch('/tasklog/api/trigger-demo/', {
+                const response = await fetch('/demo/api/trigger-demo/', {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json',
